requirements.txt
streamlit
pandas
pillow
openpyxl
pymongo
google-generativeai
dnspython
streamlit
pandas
pillow
pymongo
